nit: moving the method into a common class from the same module
